Move CSV functionality to separate package #67
