+++ Source/QuantityTypes.Csv.Tests/Test1.csv
@@ -1,0 +1,3 @@
+﻿Text,Number,Length [km]
+One,-3,1
+Two,17,4.2195E1
