+++ Source/QuantityTypes.Csv.Tests/project.json
@@ -1,0 +1,16 @@
+{
+  "version": "1.0.0-*",
+
+  "dependencies": {
+    "NETStandard.Library": "1.6.0",
+    "NUnit": "3.5.0",
+    "QuantityTypes": "0.9.0",
+    "QuantityTypes.Csv": "1.0.0-*"
+  },
+
+  "frameworks": {
+    "netstandard1.6": {
+      "imports": "dnxcore50"
+    }
+  }
+}
