+++ Source/QuantityTypes.Tests/QuantityTypes.Tests.csproj
@@ -80,10 +80,6 @@
     <Compile Include="Example.cs" />
     <Compile Include="JsonTests\ServiceStackJsonTests.cs" />
     <Compile Include="JsonTests\TestObject.cs" />
-    <Compile Include="Csv\CsvFileTests.cs" />
-    <Compile Include="Csv\CsvWriterTests.cs" />
-    <Compile Include="Csv\CsvParserTests.cs" />
-    <Compile Include="Csv\CsvTests.cs" />
     <Compile Include="Quantities\EnergyTests.cs" />
     <Compile Include="Quantities\TimeTests.cs" />
     <Compile Include="Quantities\TypographicLengthTests.cs" />
@@ -91,10 +87,10 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
-    <None Include="Csv\Test1.csv" />
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
