+++ Source/QuantityTypes.sln
@@ -26,6 +26,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QuantityTypes.Tests", "QuantityTypes.Tests\QuantityTypes.Tests.csproj", "{234E25BD-C3D0-465A-B763-69CCF5036CE9}"
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "QuantityTypes.Csv", "QuantityTypes.Csv\QuantityTypes.Csv.xproj", "{1E0F1CA8-CF62-4B65-BA66-E22F15CA70D4}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "QuantityTypes.Csv.Tests", "QuantityTypes.Csv.Tests\QuantityTypes.Csv.Tests.xproj", "{C4CD2F7B-9332-4BF8-9EED-0156BF95DF6F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -44,6 +48,14 @@ Global
 		{234E25BD-C3D0-465A-B763-69CCF5036CE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{234E25BD-C3D0-465A-B763-69CCF5036CE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{234E25BD-C3D0-465A-B763-69CCF5036CE9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1E0F1CA8-CF62-4B65-BA66-E22F15CA70D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1E0F1CA8-CF62-4B65-BA66-E22F15CA70D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1E0F1CA8-CF62-4B65-BA66-E22F15CA70D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1E0F1CA8-CF62-4B65-BA66-E22F15CA70D4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C4CD2F7B-9332-4BF8-9EED-0156BF95DF6F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C4CD2F7B-9332-4BF8-9EED-0156BF95DF6F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C4CD2F7B-9332-4BF8-9EED-0156BF95DF6F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C4CD2F7B-9332-4BF8-9EED-0156BF95DF6F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
