+++ appveyor.yml
@@ -15,6 +15,7 @@ build:
 
 after_build:
   - NuGet pack Source\QuantityTypes\QuantityTypes.nuspec -version "%GitVersion_NuGetVersion%"
+  - NuGet pack Source\QuantityTypes.Csv\QuantityTypes.Csv.nuspec -version "%GitVersion_NuGetVersion%"
 
 artifacts:
   - path: Source/QuantityTypes/bin/Release
